+++ packages/table/src/plugins/mergeCells/index.ts
@@ -1,3 +1,4 @@
 export * from './useMergeCells'
 export * from './types'
 export * from './useHeaderMergeCells'
+export * from './state'