+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @grid-table/core
 
+## 0.6.64
+
 ## 0.6.63
 
 ### Patch Changes